Restructure module tree (breaking; least invasive)
Make the `types` module `pub(crate)` and re-export its
contents through a new module structure. The new structure
is intended to better group related concepts to make
documentation easier to navigate. As a side effect, it also
hoists all definitions up a level, making the tree wider,
with individual items easier to navigate to.

While the backing types module remains substantively unchanged,
this does break the paths through which items are accessed in
client code (see test diffs for examples). Version bumped to
0.3.0 to reflect this.
