+++ cryptoki/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cryptoki"
-version = "0.2.0"
+version = "0.3.0"
 authors = ["Contributors to the Parsec project"]
 edition = '2018'
 description = "Rust-native wrapper around the PKCS #11 API"
